+++ src/include_impl/camera.cpp
@@ -24,11 +24,11 @@ namespace salmon {
 
 Camera::Camera(internal::Camera& impl) : m_impl{&impl} {}
 
-int Camera::get_x() const {return m_impl->x();}
-int Camera::get_y() const {return m_impl->y();}
+float Camera::get_x() const {return m_impl->x();}
+float Camera::get_y() const {return m_impl->y();}
 
-void Camera::set_x(int x) {m_impl->set_x(x);}
-void Camera::set_y(int y) {m_impl->set_y(y);}
+void Camera::set_x(float x) {m_impl->set_x(x);}
+void Camera::set_y(float y) {m_impl->set_y(y);}
 
 void Camera::bind_actor(Actor actor) {m_impl->bind_actor(actor.m_impl);}
 void Camera::unbind_actor() {m_impl->unbind_actor();}
